Some renaming and restructuring
