+++ precompiles/peaq-rbac/src/structs.rs
@@ -1,0 +1,15 @@
+use precompile_utils::{prelude::UnboundedBytes, EvmData};
+use sp_core::H256;
+
+#[derive(EvmData)]
+pub struct Entity {
+	pub id: H256,
+	pub name: UnboundedBytes,
+	pub enabled: bool,
+}
+
+#[derive(EvmData)]
+pub struct Role2User {
+	pub role: H256,
+	pub user: H256,
+}
